src/app/auth/signup/signup.component.ts
import { Component, OnInit, ViewChild } from '@angular/core';
import { NgForm } from '@angular/forms';
import { AuthService } from './../auth.service';

@Component({
  selector: 'app-signup',
  templateUrl: './signup.component.html',
  styleUrls: ['./signup.component.scss']
})
export class SignupComponent implements OnInit {

  @ViewChild('f') form: NgForm;

  constructor(private authService: AuthService) { }

  ngOnInit() {
  }

  onSignUp() {
    this.authService.signupUser(this.form.value.email, this.form.value.password)
                    .then(response => console.log(response))
                    .catch(error => console.log(error));
  }

}
